fixes NPE, moves flare to own pkg
