+++ pkg/flare/archive_test.go
@@ -14,7 +14,7 @@ func TestCreateArchive(t *testing.T) {
 	config.Datadog.Set("confd_path", "./test/confd")
 	config.Datadog.Set("log_file", "./test/logs/agent.log")
 	zipFilePath := mkFilePath()
-	filePath, err := createArchive(zipFilePath)
+	filePath, err := createArchive(zipFilePath, true)
 
 	assert.Nil(t, err)
 	assert.Equal(t, zipFilePath, filePath)
@@ -33,7 +33,7 @@ func TestCreateArchiveBadConfig(t *testing.T) {
 
 	common.SetupConfig("")
 	zipFilePath := mkFilePath()
-	filePath, err := createArchive(zipFilePath)
+	filePath, err := createArchive(zipFilePath, true)
 
 	assert.Nil(t, err)
 	assert.Equal(t, zipFilePath, filePath)
