+++ pkg/flare/strip.go
