+++ pkg/flare/strip_test.go
